requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
Faker==18.9.0
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.30.0
six==1.16.0
SQLAlchemy==2.0.14
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
